extract openMarkdownFile as a separate util
