+++ src/utils/obsidian.ts
@@ -92,3 +92,33 @@ export function calculateFileDistance(
 
   return distance
 }
+
+export function openMarkdownFile(
+  app: App,
+  filePath: string,
+  startLine?: number,
+) {
+  const file = app.vault.getFileByPath(filePath)
+  if (!file) return
+
+  const existingLeaf = app.workspace
+    .getLeavesOfType('markdown')
+    .find(
+      (leaf) =>
+        leaf.view instanceof MarkdownView && leaf.view.file?.path === file.path,
+    )
+
+  if (existingLeaf) {
+    app.workspace.setActiveLeaf(existingLeaf, { focus: true })
+
+    if (startLine) {
+      const view = existingLeaf.view as MarkdownView
+      view.setEphemeralState({ line: startLine })
+    }
+  } else {
+    const leaf = app.workspace.getLeaf('tab')
+    leaf.openFile(file, {
+      eState: startLine ? { line: startLine } : undefined,
+    })
+  }
+}
